fix-2

+++ src/main/java/com/codegym/casestudymodule3/model/Product.java
@@ -74,7 +74,7 @@ public class Product {
         return product_price;
     }
 
-    public void setProduct_price(Float product_price) {
+    public void setProduct_price(double product_price) {
         this.product_price = product_price;
     }
